requirements.txt
langchain
langchain_groq
python-dotenv
langchain_core
streamlit
validators
bs4
requests
ipykernel
fastapi
uvicorn